move FacetDefinition to @gen3/core
