+++ packages/core/src/features/facets/index.ts
@@ -1,0 +1,1 @@
+export * from './types';
