+++ packages/frontend/src/components/facets/types.ts
@@ -2,11 +2,12 @@ import {
   CombineMode,
   DataFetchingResult,
   EnumFilterValue,
+  FacetDefinition,
   IndexAndField,
+  NumericFromTo,
   Operation,
 } from '@gen3/core';
 import React, { ComponentType, ReactNode } from 'react';
-import { NumericFromTo } from './NumericRangeFacet/types';
 
 export type QueryOptions = Record<string, unknown>;
 
@@ -49,20 +50,6 @@ export interface FacetCardProps<T extends FacetCommonHooks> {
     readonly iconStyle?: string; // optional facet button component
   };
 }
-
-export type FacetType =
-  | 'enum'
-  | 'exact'
-  | 'range'
-  | 'age'
-  | 'year'
-  | 'years'
-  | 'days'
-  | 'percent'
-  | 'datetime'
-  | 'toggle'
-  | 'multiselect'
-  | 'upload';
 
 // required functions
 export type ClearFacetFunction = (field: string) => void;
@@ -258,11 +245,6 @@ export interface SelectedFields {
   updateSelectedField: (facet: string) => void;
 }
 
-export interface AllowableRange {
-  readonly minimum: number;
-  readonly maximum: number;
-}
-
 export interface CohortBuilderCategoryConfig {
   readonly label: string;
   readonly facets: ReadonlyArray<string>;
@@ -279,18 +261,4 @@ export interface EnumChartProps {
   readonly maxBins: number;
   readonly height: number;
   readonly valueLabel?: string;
-}
-
-export interface FacetDefinition {
-  readonly description?: string; //description from _mapping
-  readonly field: string; // full name of field
-  readonly dataField?: string; // deprecated
-  readonly index: string; // what dataType is this facet for
-  readonly type: FacetType; // classified type based on type + name: e.g. age, year, enumeration, etc
-  readonly range?: AllowableRange; // range of value types
-  readonly hasData?: boolean; // does this facet have data
-  readonly label?: string; // label for facet
-  readonly sharedWithIndices?: Array<IndexAndField>; // if this filter is denormalized across indices
-  readonly moveValuesToBottom?: Array<string>;
-  readonly excludeValues?: Array<string>;
 }
