+++ packages/frontend/src/components/facets/utils.ts
@@ -25,7 +25,6 @@ import {
 } from '@gen3/core';
 import {
   ClearFacetFunction,
-  FacetDefinition,
   FacetSortType,
   FieldToName,
   FromToRange,
@@ -34,6 +33,7 @@ import {
 } from './types';
 import { isArray } from 'lodash';
 import { TabConfig } from '../../features/CohortBuilder/types';
+import { FacetDefinition } from '@gen3/core';
 
 export const getAllFieldsFromFilterConfigs = (
   filterTabConfigs: ReadonlyArray<TabConfig>,
