+++ packages/frontend/src/features/CohortDiscovery/FilterPanels/createFacetPanel.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
-import { FacetDefinition } from '../../../components/facets/types';
 import EnumFacetPanel, { EnumFacetPanelDataHooks } from './EnumFacetPanel';
+import { FacetDefinition } from '@gen3/core';
 
 export const createFacetPanel = (
   facet: FacetDefinition,
