+++ packages/frontend/src/features/CohortDiscovery/IndexPanel.tsx
@@ -4,8 +4,6 @@ import { Box, Center, Flex, Stack, Text, Title } from '@mantine/core';
 import FacetSelectionPanel from './FacetSelectionPanel';
 import { useFilterExpandedState, useToggleExpandFilter } from './hooks';
 
-import { FacetDefinition } from '../../components/facets/types';
-
 import { selectCurrentCohortIndexFilters } from './CohortManagment/CohortManagerSelectors';
 import { CohortDiscoveryGroup } from './types';
 import {
@@ -25,6 +23,7 @@ import {
 } from './SelectedFacetsSlice';
 import { useUnsecureRoundedAggsQuery } from './queryApi';
 import CohortQueryExpression from './CohortQueryExpression';
+import { FacetDefinition } from '@gen3/core';
 
 const IndexPanel = ({
   dataConfig,
